pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "imbalanced-learn>=0.12.4",
    "ipykernel>=6.30.1",
    "joblib>=1.5.1",
    "jupyterlab>=4.4.5",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "shap>=0.48.0",
    "xgboost>=2.1.4",
]
